test: relocate the hosted site cases to agent/hosted-site-coordinator.test.ts
Second commit of PR 2. Relocation, not a rewrite: every assertion is
byte-identical and each test still constructs a real AgentService through
agent-service-test-harness.ts, driving the fake provider and asserting through
the public approval/conversation events and the persisted marker rows. That
matters most for the two restart cases, which stop and rebuild the service to
prove a durable marker replays without repeating the deploy — as controller unit
tests they would protect nothing.

Six tests move: approval before a mutation, failed update and successful delete
as separate transitions, the durable result replayed after restart, legacy
metadata normalization, the interrupted marker after restart, and the successful
result kept when the provider response cannot be delivered.

restart.test.ts:177 stays where it is. The plan listed it as a hosted-site seed,
and that is all it is — the test is about deleteBot: it seeds one pending
terminal event, then asserts orchestration events and receipts are gone and that
deleting an agent with running work is refused. It crosses domains, so it stays
in the shard named for the behaviour.

agent-service.approvals.test.ts keeps the one case that is genuinely about the
generic approval surface: Codex approvals not auto-accepting, and one-shot
decision mapping.

Tests: agent/hosted-site-coordinator.test.ts 6 passed,
agent-service.approvals.test.ts 1 passed; the same 7 the shard ran before the
split. tsc on tsconfig.node.json and biome on both files clean.
